requirements.txt
python-dateutil
tzlocal
pytz
requests
schwab-py
keyboard
pandas>=1.5.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv
textual
python-telegram-bot>=22.3,<23
